pyproject.toml
[project]
name = "cupertino"
version = "0.1.0"
description = "Zone monitoring system with YOLO and MQTT"
requires-python = ">=3.12.12"
dependencies = [
    "inference>=0.58.3",
    "paho-mqtt>=1.6.1,<2.0",
    "pyyaml>=6.0",
    "supervision>=0.26.1",
    "tk>=0.1.0",
    "ultralytics>=8.3.220",
]

[tool.setuptools]
packages = ["cupertino_zone", "cupertino_mqtt", "cupertino_processor", "cupertino_control", "cupertino_cli"]

[project.scripts]
cupertino-cli = "cupertino_cli.cli:main"
